fix: test cases for default endpoint provider and acc uploading

+++ test/form_up.test.js
@@ -9,8 +9,7 @@ const qiniu = require('../index.js');
 const {
     Endpoint,
     Region,
-    SERVICE_NAME,
-    StaticRegionsProvider
+    SERVICE_NAME
 } = qiniu.httpc;
 
 const {
@@ -67,6 +66,10 @@ describe('test form up', function () {
     // delete all the files uploaded
     const keysToDelete = [];
     after(function () {
+        if (!keysToDelete.length) {
+            return;
+        }
+
         const deleteOps = keysToDelete.map(k =>
             qiniu.rs.deleteOp(bucketName, k)
         );
@@ -448,13 +451,31 @@ describe('test form up', function () {
 
     describe('test form up#accelerateUploading', function () {
         const accConfig = new qiniu.conf.Config();
-        // accConfig.useHttpsDomain = true;
+        accConfig.useHttpsDomain = true;
         accConfig.accelerateUploading = true;
-        const accFormUploader = new qiniu.form_up.FormUploader(accConfig);
         const bucketNameWithoutAcc = 'bucket-without-acc-' + Math.floor(Math.random() * 100000);
+        const accPutPolicy = new qiniu.rs.PutPolicy({
+            scope: bucketNameWithoutAcc
+        });
+        const accUploadToken = accPutPolicy.uploadToken(mac);
+        const accFormUploader = new qiniu.form_up.FormUploader(accConfig);
+        const accKeysToDelete = [];
 
         before(function () {
             return bucketManager.createBucket(bucketNameWithoutAcc);
+        });
+
+        after(function () {
+            if (!accKeysToDelete.length) {
+                return;
+            }
+            return bucketManager.batch(accKeysToDelete.map(k => qiniu.rs.deleteOp(bucketNameWithoutAcc, k)))
+                .then(({ data, resp }) => {
+                    if (!Array.isArray(data)) {
+                        console.log(resp);
+                    }
+                    return bucketManager.deleteBucket(bucketNameWithoutAcc);
+                });
         });
 
         it('upload acc normally', function () {
@@ -464,7 +485,9 @@ describe('test form up', function () {
                 accFormUploader.putFile(uploadToken, key, testFilePath2, putExtra, callback)
             );
 
-            const checkFunc = ({ data }) => {
+            const checkFunc = ({ data, resp }) => {
+                const isAccelerateUploading = (resp.requestUrls || []).some(url => url.includes('kodo-accelerate'));
+                should.ok(isAccelerateUploading, `should using acc host, but requestUrls: ${JSON.stringify(resp.requestUrls)}`);
                 data.should.have.keys('key', 'hash');
             };
 
@@ -480,10 +503,6 @@ describe('test form up', function () {
         it('upload acc unavailable fallback to src', function () {
             const key = 'storage_putFile_acc_test' + Math.floor(Math.random() * 100000);
 
-            const putPolicy = new qiniu.rs.PutPolicy({
-                scope: bucketNameWithoutAcc
-            });
-            const uploadToken = putPolicy.uploadToken(mac);
             const r1 = Region.fromRegionId('z0');
             r1.services[SERVICE_NAME.UP_ACC] = [
                 new Endpoint(`${bucketNameWithoutAcc}.kodo-accelerate.cn-east-1.qiniucs.com`),
@@ -492,7 +511,7 @@ describe('test form up', function () {
             accConfig.regionsProvider = r1;
 
             const promises = doAndWrapResultPromises(callback =>
-                accFormUploader.putFile(uploadToken, key, testFilePath2, putExtra, callback)
+                accFormUploader.putFile(accUploadToken, key, testFilePath2, putExtra, callback)
             );
 
             const checkFunc = ({ data }) => {
@@ -504,17 +523,13 @@ describe('test form up', function () {
                 .then(() => promises.native)
                 .then(checkFunc)
                 .then(() => {
-                    keysToDelete.push(key);
+                    accKeysToDelete.push(key);
                 });
         });
 
         it('upload acc network error fallback to src', function () {
             const key = 'storage_putFile_acc_test' + Math.floor(Math.random() * 100000);
 
-            const putPolicy = new qiniu.rs.PutPolicy({
-                scope: bucketNameWithoutAcc
-            });
-            const uploadToken = putPolicy.uploadToken(mac);
             const r1 = Region.fromRegionId('z0');
             r1.services[SERVICE_NAME.UP_ACC] = [
                 new Endpoint('qiniu-acc.fake.qiniu.com'),
@@ -523,7 +538,7 @@ describe('test form up', function () {
             accConfig.regionsProvider = r1;
 
             const promises = doAndWrapResultPromises(callback =>
-                accFormUploader.putFile(uploadToken, key, testFilePath2, putExtra)
+                accFormUploader.putFile(accUploadToken, key, testFilePath2, putExtra)
             );
 
             const checkFunc = ({ data }) => {
@@ -535,7 +550,7 @@ describe('test form up', function () {
                 .then(() => promises.native)
                 .then(checkFunc)
                 .then(() => {
-                    keysToDelete.push(key);
+                    accKeysToDelete.push(key);
                 });
         });
     });
